+++ src/pages/Logs.tsx
@@ -17,10 +17,15 @@ export class LogsPage extends BasePage {
     scene: undefined as EmbeddedScene | undefined,
   };
 
+  private mounted = true;
+
   constructor(props: any) {
     super(props);
 
     this.prepareDs().then(() => {
+      if (!this.mounted) {
+        return;
+      }
       const jsonData = this.cluster?.instanceSettings.jsonData as KubegrafDSOptions | undefined;
       const logsUid = jsonData?.logs_uid;
       this.setState({
@@ -32,8 +37,14 @@ export class LogsPage extends BasePage {
     });
 
     this.getAvailableClusters().then((res) => {
-      this.setState({ clusters: res });
+      if (this.mounted) {
+        this.setState({ clusters: res });
+      }
     });
+  }
+
+  componentWillUnmount() {
+    this.mounted = false;
   }
 
   render() {